Clean old chat messages
It has been reported by CN community that the game will run out of
memory after the user keeps in a lobby with a large amount of messages
received. This commit, of which the idea was from @Tstar00, removes too
old messages to avoid them from running out of the available 2GB memory
space.